creating seprate app for rights
