+++ project_electron/orgs/templates/orgs/update.html
@@ -66,7 +66,7 @@
 			<div class="box-header with-border">
 				<h3 class="box-title">Rights Statements</h3>
 			</div>
-			<div class="box-body">
+			<div class="box-body no-padding">
 				{% include 'parts/rights_table.html' %}
 			</div>
 		</div>
