+++ project_electron/orgs/templates/parts/rights_table.html
@@ -1,4 +1,4 @@
-<table class="table table-bordered dataTable">
+<table class="table table-striped">
   <thead>
     <tr>
       <th>Record Type</th>
