+++ project_electron/orgs/urls.py
@@ -8,6 +8,6 @@ urlpatterns = [
     url(r'^$', 	OrganizationListView.as_view(), name='orgs-list'),
     url(r'^(?P<pk>\d+)/edit/$', OrganizationEditView.as_view(), name='orgs-edit'),
     url(r'^(?P<pk>\d+)/transfers/$', OrganizationTransfersView.as_view(), name='orgs-transfers-report'),
-    url(r'^(?P<pk>\d+)/rights/', include('orgs.rights_urls')),
+    url(r'^(?P<pk>\d+)/rights/', include('rights.urls')),
 
 ]
